Avoid redundant downloads in some cases
Given how we generate filenames from metadata so
far, when we are given a filepath without an
extension, it is not possible to have a different
file with the same name. This is helpful to leverage
so we can avoid having to download the file in order
to determine its extension before we can properly
check for the file existence.

It may be that this will not hold in the future
in which case we would have to fallback to reading
mime types from metadata (reverted in 1fda654a).
Note though that we can still have incorrect mime
types recorded in NFT metadata (I have seen this in
at least one case) so we should always depend on
extension detection in order to determine the final
file extension.